+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -26,11 +26,6 @@ public class UserDao {
         jdbcTemplate.update(sql, user.getAccount(), user.getPassword(), user.getEmail());
     }
 
-//    public void update(final Connection conn, final User user) {
-//        final String sql = "update users set account = ?, password = ?, email = ? where id = ?";
-//        jdbcTemplate.update(conn, sql, user.getAccount(), user.getPassword(), user.getEmail(), user.getId());
-//    }
-
     public void update(final User user) {
         final String sql = "update users set account = ?, password = ?, email = ? where id = ?";
         jdbcTemplate.update(sql, user.getAccount(), user.getPassword(), user.getEmail(), user.getId());
